Remove isLarge from Buttons (#74)
The Button core component no longer supports isLarge:
https://github.com/WordPress/gutenberg/commit/7733aea1f922c8b522a052c4e6123b712911480d#diff-1204d0def76e28ff29950465cc842e3fafcee022c7e97c76fa5cf5baa4c4505a

This means that several parts of the code (removed in this commit) throw
a console error about being React being unable to add `isLarge` to a DOM
element. Removing these references removes the error, and doesn't seem
to have any effect on the operation of the plugin.