Correctly filter by post type when searching posts
Setting the `type` filter based on the value passed to the onChange
callback sets the filter to an array of `{ label, value }` objects.
When these are passed in the API request, they are not recognized as
a valid value for the "type" parameter, and are stripped. We need to
set the filter to an array of only the `value` parameter so that
queries can be correctly filtered by post type.